+++ src/test/java/org/embulk/output/elasticsearch/TestOpenSearchOutputPlugin.java
@@ -22,14 +22,15 @@ import org.embulk.config.ConfigException;
 import org.embulk.config.ConfigSource;
 import org.embulk.config.TaskReport;
 import org.embulk.config.TaskSource;
-import org.embulk.output.elasticsearch.ElasticsearchOutputPluginDelegate.AuthMethod;
-import org.embulk.output.elasticsearch.ElasticsearchOutputPluginDelegate.Mode;
-import org.embulk.output.elasticsearch.ElasticsearchOutputPluginDelegate.PluginTask;
+import org.embulk.output.elasticsearch.OpenSearchOutputPluginDelegate.AuthMethod;
+import org.embulk.output.elasticsearch.OpenSearchOutputPluginDelegate.Mode;
+import org.embulk.output.elasticsearch.OpenSearchOutputPluginDelegate.PluginTask;
 import org.embulk.spi.OutputPlugin;
 import org.embulk.spi.Page;
 import org.embulk.spi.PageTestUtils;
 import org.embulk.spi.Schema;
 import org.embulk.spi.TransactionalPageOutput;
+import org.embulk.standards.CsvParserPlugin;
 import org.embulk.util.config.ConfigMapper;
 import org.embulk.util.config.ConfigMapperFactory;
 import org.junit.After;
@@ -40,34 +41,40 @@ import org.opensearch.client.opensearch.OpenSearchClient;
 import org.opensearch.client.opensearch.core.SearchResponse;
 
 import java.io.IOException;
+import java.time.Instant;
 import java.util.Arrays;
 import java.util.List;
 
-import static org.embulk.output.elasticsearch.ElasticsearchTestUtils.ES_INDEX;
+import static org.embulk.output.elasticsearch.OpenSearchTestUtils.ES_BULK_ACTIONS;
+import static org.embulk.output.elasticsearch.OpenSearchTestUtils.ES_BULK_SIZE;
+import static org.embulk.output.elasticsearch.OpenSearchTestUtils.ES_CONCURRENT_REQUESTS;
+import static org.embulk.output.elasticsearch.OpenSearchTestUtils.ES_ID;
+import static org.embulk.output.elasticsearch.OpenSearchTestUtils.ES_INDEX;
+import static org.embulk.output.elasticsearch.OpenSearchTestUtils.ES_NODES;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.junit.Assert.assertTrue;
 
-public class TestElasticsearchOutputPluginJSON
+public class TestOpenSearchOutputPlugin
 {
-    private static final ConfigMapperFactory CONFIG_MAPPER_FACTORY = ElasticsearchOutputPlugin.CONFIG_MAPPER_FACTORY;
-    private static final ConfigMapper CONFIG_MAPPER = ElasticsearchOutputPlugin.CONFIG_MAPPER;
+    private static final ConfigMapperFactory CONFIG_MAPPER_FACTORY = OpenSearchOutputPlugin.CONFIG_MAPPER_FACTORY;
+    private static final ConfigMapper CONFIG_MAPPER = OpenSearchOutputPlugin.CONFIG_MAPPER;
 
     @Rule
     public EmbulkTestRuntime runtime = new EmbulkTestRuntime();
-    private ElasticsearchOutputPlugin plugin;
-    private ElasticsearchTestUtils utils;
+    private OpenSearchOutputPlugin plugin;
+    private OpenSearchTestUtils utils;
     private OpenSearchClient openSearchClient;
 
     @Before
     public void createResources() throws Exception
     {
-        utils = new ElasticsearchTestUtils();
+        utils = new OpenSearchTestUtils();
         utils.initializeConstant();
-        final PluginTask task = CONFIG_MAPPER.map(utils.configJSON(), PluginTask.class);
+        final PluginTask task = CONFIG_MAPPER.map(utils.config(), PluginTask.class);
         utils.prepareBeforeTest(task);
 
-        plugin = new ElasticsearchOutputPlugin();
+        plugin = new OpenSearchOutputPlugin();
 
         openSearchClient = utils.client();
     }
@@ -86,15 +93,48 @@ public class TestElasticsearchOutputPluginJSON
     @Test
     public void testDefaultValues()
     {
-        final PluginTask task = CONFIG_MAPPER.map(utils.configJSON(), PluginTask.class);
+        final PluginTask task = CONFIG_MAPPER.map(utils.config(), PluginTask.class);
         assertThat(task.getIndex(), is(ES_INDEX));
+    }
+
+    @Test
+    public void testDefaultValuesNull()
+    {
+        final ConfigSource config = runtime.getExec().newConfigSource()
+            .set("in", utils.inputConfig())
+            .set("parser", utils.parserConfig(utils.schemaConfig()))
+            .set("type", "elasticsearch")
+            .set("mode", "") // NULL
+            .set("nodes", ES_NODES)
+            .set("index", ES_INDEX)
+            .set("id", ES_ID)
+            .set("bulk_actions", ES_BULK_ACTIONS)
+            .set("bulk_size", ES_BULK_SIZE)
+            .set("concurrent_requests", ES_CONCURRENT_REQUESTS
+            );
+        Schema schema = utils.oldParserConfig(runtime).loadConfig(CsvParserPlugin.PluginTask.class).getSchemaConfig().toSchema();
+        try {
+            plugin.transaction(config, schema, 0, new OutputPlugin.Control()
+            {
+                @Override
+                public List<TaskReport> run(TaskSource taskSource)
+                {
+                    return Lists.newArrayList(CONFIG_MAPPER_FACTORY.newTaskReport());
+                }
+            });
+        }
+        catch (Throwable t) {
+            if (t instanceof RuntimeException) {
+                assertTrue(t instanceof ConfigException);
+            }
+        }
     }
 
     @Test
     public void testTransaction()
     {
-        ConfigSource config = utils.configJSON();
-        Schema schema = utils.JSONSchema();
+        ConfigSource config = utils.config();
+        Schema schema = utils.oldParserConfig(runtime).loadConfig(CsvParserPlugin.PluginTask.class).getSchemaConfig().toSchema();
         plugin.transaction(config, schema, 0, new OutputPlugin.Control()
         {
             @Override
@@ -109,8 +149,8 @@ public class TestElasticsearchOutputPluginJSON
     @Test
     public void testResume()
     {
-        ConfigSource config = utils.configJSON();
-        Schema schema = utils.JSONSchema();
+        ConfigSource config = utils.config();
+        Schema schema = utils.oldParserConfig(runtime).loadConfig(CsvParserPlugin.PluginTask.class).getSchemaConfig().toSchema();
         final PluginTask task = CONFIG_MAPPER.map(config, PluginTask.class);
         plugin.resume(task.toTaskSource(), schema, 0, new OutputPlugin.Control()
         {
@@ -125,8 +165,8 @@ public class TestElasticsearchOutputPluginJSON
     @Test
     public void testCleanup()
     {
-        ConfigSource config = utils.configJSON();
-        Schema schema = utils.JSONSchema();
+        ConfigSource config = utils.config();
+        Schema schema = utils.oldParserConfig(runtime).loadConfig(CsvParserPlugin.PluginTask.class).getSchemaConfig().toSchema();
         final PluginTask task = CONFIG_MAPPER.map(config, PluginTask.class);
         plugin.cleanup(task.toTaskSource(), schema, 0, Arrays.asList(CONFIG_MAPPER_FACTORY.newTaskReport()));
         // no error happens
@@ -135,8 +175,8 @@ public class TestElasticsearchOutputPluginJSON
     @Test
     public void testOutputByOpen() throws Exception
     {
-        ConfigSource config = utils.configJSON();
-        Schema schema = utils.JSONSchema();
+        ConfigSource config = utils.config();
+        Schema schema = utils.oldParserConfig(runtime).loadConfig(CsvParserPlugin.PluginTask.class).getSchemaConfig().toSchema();
         final PluginTask task = CONFIG_MAPPER.map(config, PluginTask.class);
         plugin.transaction(config, schema, 0, new OutputPlugin.Control() {
             @Override
@@ -147,7 +187,7 @@ public class TestElasticsearchOutputPluginJSON
         });
         TransactionalPageOutput output = plugin.open(task.toTaskSource(), schema, 0);
 
-        List<Page> pages = PageTestUtils.buildPage(runtime.getBufferAllocator(), schema, 1L, 32864L, "2015-01-27 19:23:49", "2015-01-27",  true, 123.45, "embulk");
+        List<Page> pages = PageTestUtils.buildPage(runtime.getBufferAllocator(), schema, 1L, 32864L, Instant.ofEpochSecond(1422386629), Instant.ofEpochSecond(1422316800),  true, 123.45, "embulk");
         assertThat(pages.size(), is(1));
         for (Page page : pages) {
             output.add(page);
@@ -155,7 +195,7 @@ public class TestElasticsearchOutputPluginJSON
 
         output.finish();
         output.commit();
-        Thread.sleep(1500); // Need to wait until index done
+        Thread.sleep(3000); // Need to wait until index done
 
         SearchResponse<IndexData> response = openSearchClient.search(s -> s.index(ES_INDEX), IndexData.class);
 
@@ -166,49 +206,8 @@ public class TestElasticsearchOutputPluginJSON
         IndexData record = response.hits().hits().get(0).source();
         assertThat(record.getId(), is(1L));
         assertThat(record.getAccount(), is(32864L));
-        assertThat(record.getTime(), is("2015-01-27 19:23:49"));
-        assertThat(record.getPurchase(), is("2015-01-27"));
-        assertThat(record.getFlg(), is(true));
-        assertThat(record.getScore(), is(123.45));
-        assertThat(record.getComment(), is("embulk"));
-    }
-
-    @Test
-    public void testOutputByOpenWithNulls() throws Exception
-    {
-        ConfigSource config = utils.configJSON();
-        Schema schema = utils.JSONSchema();
-        final PluginTask task = CONFIG_MAPPER.map(config, PluginTask.class);
-        plugin.transaction(config, schema, 0, new OutputPlugin.Control() {
-            @Override
-            public List<TaskReport> run(TaskSource taskSource)
-            {
-                return Lists.newArrayList(CONFIG_MAPPER_FACTORY.newTaskReport());
-            }
-        });
-        TransactionalPageOutput output = plugin.open(task.toTaskSource(), schema, 0);
-
-        List<Page> pages = PageTestUtils.buildPage(runtime.getBufferAllocator(), schema, 2L, null, null, "2015-01-27",  true, 123.45, "embulk");
-        assertThat(pages.size(), is(1));
-        for (Page page : pages) {
-            output.add(page);
-        }
-
-        output.finish();
-        output.commit();
-        Thread.sleep(1500); // Need to wait until index done
-
-        SearchResponse<IndexData> response = openSearchClient.search(s -> s.index(ES_INDEX), IndexData.class);
-
-        int totalHits = response.hits().hits().size();
-
-        assertThat(totalHits, is(1));
-
-        IndexData record = response.hits().hits().get(0).source();
-        assertThat(record.getId(), is(2L));
-        assertTrue(record.getAccount() == null);
-        assertTrue(record.getTime() == null);
-        assertThat(record.getPurchase(), is("2015-01-27"));
+        assertThat(record.getTime(), is("2015-01-27T19:23:49.000+0000"));
+        assertThat(record.getPurchase(), is("2015-01-27T00:00:00.000+0000"));
         assertThat(record.getFlg(), is(true));
         assertThat(record.getScore(), is(123.45));
         assertThat(record.getComment(), is("embulk"));
@@ -217,8 +216,8 @@ public class TestElasticsearchOutputPluginJSON
     @Test
     public void testOpenAbort()
     {
-        ConfigSource config = utils.configJSON();
-        Schema schema = utils.JSONSchema();
+        ConfigSource config = utils.config();
+        Schema schema = utils.oldParserConfig(runtime).loadConfig(CsvParserPlugin.PluginTask.class).getSchemaConfig().toSchema();
         final PluginTask task = CONFIG_MAPPER.map(config, PluginTask.class);
         TransactionalPageOutput output = plugin.open(task.toTaskSource(), schema, 0);
         output.abort();
